pyproject.toml
[project]
name = "drug-shortage"
version = "0.1.0"
description = "Drug shortage dashboard data pipeline"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "python-dateutil>=2.8.2",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "sqlalchemy>=2.0.43",
    "psycopg2-binary>=2.9.10",
    "dbt-postgres>=1.9.0",
    "ipykernel>=6.30.1",
    "dash>=3.2.0",
    "plotly>=6.3.0",
    "streamlit>=1.49.1",
]
